Minor improvements and major refactoring of editor file browser
Do not show a confirmation popup anymore when saving a map to its current name with the Save button in the File menu. This makes the behavior consistent with the Ctrl+S behavior, which also does not show a confirmation. This is also consistent with the behavior of other applications, that do not ask whether to overwrite the existing file when using the "Save" or Ctrl+S action.

Fix location of maps being opened from folders inside of links not being detected correctly.

Selected newly created folder automatically.

Red buttons for delete and overwrite popup buttons. Large popup for delete.

Fix parent folder list entry being missing in sub-folders if there is only one storage location.

Minor improvement of file browser UI layout.

Move all variables and functions related to the editor file browser from `CEditor` class to a separate class `CFileBrowser` in separate files. Encapsulate.